Concluido lógica de recorte de imagem e refatoração de códigos

+++ src/com/maroni/util/util/DimensoesImagem.java
@@ -5,96 +5,120 @@ import java.io.Serializable;
 public class DimensoesImagem implements Serializable {
 	private static final long serialVersionUID = 1631872295227745456L;
 	
-	private double height;
-	private double width;
-	private double x1;
-	private double x2;
-	private double y1;
-	private double y2;
+	private int height;
+	private int width;
+	private int x1;
+	private int x2;
+	private int y1;
+	private int y2;
 
 	/**
 	 * @return the height
 	 */
-	public double getHeight() {
+	public int getHeight() {
 		return height;
 	}
+
+
 
 	/**
 	 * @param height the height to set
 	 */
-	public void setHeight(double height) {
+	public void setHeight(int height) {
 		this.height = height;
 	}
+
+
 
 	/**
 	 * @return the width
 	 */
-	public double getWidth() {
+	public int getWidth() {
 		return width;
 	}
+
+
 
 	/**
 	 * @param width the width to set
 	 */
-	public void setWidth(double width) {
+	public void setWidth(int width) {
 		this.width = width;
 	}
+
+
 
 	/**
 	 * @return the x1
 	 */
-	public double getX1() {
+	public int getX1() {
 		return x1;
 	}
+
+
 
 	/**
 	 * @param x1 the x1 to set
 	 */
-	public void setX1(double x1) {
+	public void setX1(int x1) {
 		this.x1 = x1;
 	}
+
+
 
 	/**
 	 * @return the x2
 	 */
-	public double getX2() {
+	public int getX2() {
 		return x2;
 	}
+
+
 
 	/**
 	 * @param x2 the x2 to set
 	 */
-	public void setX2(double x2) {
+	public void setX2(int x2) {
 		this.x2 = x2;
 	}
+
+
 
 	/**
 	 * @return the y1
 	 */
-	public double getY1() {
+	public int getY1() {
 		return y1;
 	}
+
+
 
 	/**
 	 * @param y1 the y1 to set
 	 */
-	public void setY1(double y1) {
+	public void setY1(int y1) {
 		this.y1 = y1;
 	}
+
+
 
 	/**
 	 * @return the y2
 	 */
-	public double getY2() {
+	public int getY2() {
 		return y2;
 	}
+
+
 
 	/**
 	 * @param y2 the y2 to set
 	 */
-	public void setY2(double y2) {
+	public void setY2(int y2) {
 		this.y2 = y2;
 	}
+
+
 
 	@Override
 	public String toString() {
